# Conflicts:
#	infrastructure/docker-compose.yml
#	test.http

+++ gameStatisticsContext/src/main/java/be/kdg/prog6/gameStatisticsContext/adapter/in/AdminStatisticsDto.java
@@ -6,7 +6,7 @@ public record AdminStatisticsDto(
         UUID gameId,
         int totalScore,
         int totalGamesPlayed,
-        long totalTimePlayed,
+        double totalTimePlayed,
         int highestScore,
         int movesMade,
         double averageGameDuration,
